+++ src/ui/MenuState.java
@@ -47,7 +47,7 @@ public class MenuState implements IGameState {
   }
 
   @Override
-  public void update(final GameStateManager stateGame, final int delta) {
+  public void update(final GameStateManager stateGame, final float delta) {
     if (Keyboard.isKeyDown(Keyboard.KEY_F2)) {
       stateGame.quit();
     } else if (Keyboard.isKeyDown(Keyboard.KEY_UP)) {
